Split derive macros
